Move git SHA fetcher into its own file.
Fix tests.
