+++ pydis_site/settings.py
@@ -20,7 +20,7 @@ import sentry_sdk
 from django.contrib.messages import constants as messages
 from sentry_sdk.integrations.django import DjangoIntegration
 
-from pydis_site.utils.resources import get_git_sha
+from pydis_site.utils import get_git_sha
 
 if typing.TYPE_CHECKING:
     from django.contrib.auth.models import User
