+++ pydis_site/utils/__init__.py
@@ -1,0 +1,3 @@
+from .utils import get_git_sha
+
+__all__ = ['get_git_sha']
